+++ libdariadb/utils/crc.h
@@ -1,9 +1,10 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 
 namespace dariadb {
 namespace utils {
-size_t crc32(const void *buffer, const size_t size);
+uint32_t crc32(const void *buffer, const size_t size);
 }
 }
